MI5-41 Refaktoriranje naziva metoda, izdvajanje elemenata u zasebne datoteke
