+++ core/src/main/java/hr/foi/air/core/transactions/TransactionsSuccessOutcomeListener.kt
@@ -1,11 +1,11 @@
 package hr.foi.air.core.transactions
 
 interface TransactionsSuccessOutcomeListener {
-    fun onSuccessfulTransactionsFetch(
+    fun onSuccessfulTransactionsSuccessFetch(
         totalTransactions: Int,
         successfulTransactions: Int,
         rejectedTransactions: Int,
         canceledTransactions: Int
     )
-    fun onFailedTransactionsFetch(failureMessage: String)
+    fun onFailedTransactionsSuccessFetch(failureMessage: String)
 }
